complted

+++ backend2.0/miniProject/routes/notes.route.js
@@ -1,6 +1,7 @@
 const express=require("express") ;
 const {notesModel}=require("../models/notes.model") ;
 const notesRoute=express.Router() ;
+
 notesRoute.get("/",async(req,res)=>{
     try{
         const notes=await notesModel.find() ;
@@ -15,10 +16,11 @@ notesRoute.get("/",async(req,res)=>{
 //create notes
 notesRoute.post("/create" ,async(req,res)=>{
     const data=req.body ;
-    console.log(data ,"body")
+    //console.log(data ,"body")
     try{
+        
         const note= new notesModel(data) ;
-        console.log(note ,"no")
+        //console.log(note ,"no")
         await note.save() ;
         res.send("post created")
     }catch(err){
@@ -32,38 +34,39 @@ res.send("something sent wrong 500")
 notesRoute.patch("/update/:id" ,async(req,res)=>{
     const id=req.params.id ;
     const data=req.body ;
-    console.log(id ,data)
+    //console.log(id ,data)
     try{
-await notesModel.findByIdAndUpdate(id, data) ;
-res.status(404).send({statusCode:"200" ,mesg:"updated successfully"})
+        //compare note id and user id
+        const userID_note=await notesModel.findOne({"_id":id}) ;
+        const userID=req.body.userID ;
+        console.log(req.body)
+        if(userID===userID_note.userID){
+            await notesModel.findByIdAndUpdate(id, data) ;
+            res.status(200).send({statusCode:"200" ,mesg:"updated successfully"})
+        }else{
+            res.send("not authoried")
+        }
+
     }catch(err){
         console.log(err)
     }
 })
 
-notesRoute.put("/update/:id", async (req, res) => {
-    const id = req.params.id;
-    const data = req.body;
-    try {
-        const updatedNote = await notesModel.findByIdAndUpdate(id, data, { new: true });
-        if (!updatedNote) {
-            return res.status(404).send("Note not found");
-        }
-        res.send("Note updated");
-    } catch (err) {
-        console.error(err);
-        res.status(500).send("Something went wrong");
-    }
-});
+
 
 //delete notes
 notesRoute.delete("/delete/:id" ,async(req,res)=>{
     const id=req.params.id ;
-   console.log(id)
+   
     try{
-await notesModel.findByIdAndDelete(id) ;
+        const userID_note=await notesModel.findOne({"_id":id}) ;
+        const userID=req.body.userID ;
+        if(userID_note===userID){
+            await notesModel.findByIdAndDelete(id) ;
  
-res.status(200).send({ statusCode: 200, message: "Note deleted" });
+            res.status(200).send({ statusCode: 200, message: "Note deleted" });
+        }
+
 
     }catch(err){
         console.log(err)
